requirements.txt
openai
tweepy
python-dotenv
requests
jinja2
playwright
Pillow